move everything related to the kanban state to 'state.tsx'
