+++ libs/kanban/src/lib/kanban.tsx
@@ -15,81 +15,26 @@ import {
 } from "react-beautiful-dnd";
 import {useTheme} from "@emotion/react";
 // import '@theming/theme'; // TODO this would break the UI tests because jest/babel is not configured correctly. workaround is an ignored import on the next line
-import {lightTheme as ignoredValue} from '@theming/theme';
-import {createInitialState, KanbanItem, KanbanList, KanbanState} from "./data";
+import {lightTheme as ignoredValue} from '@theming/theme'; // workaround (as mentioned above)
+import {createInitialState, KanbanAction, KanbanItem, KanbanList, KanbanState, stateReducer} from "./state";
 
-type KanbanAddItemAction = {
-  type: "addItem",
-  targetListId: string,
-  item: KanbanItem,
-}
-
-type KanbanUpdateItemAction = {
-  type: "updateItem",
-  itemId: string,
-  checked: boolean,
-}
-
-type KanbanMoveItemAction = {
-  type: "moveItem",
-  itemId: string,
-  sourceListId: string,
-  sourceIndex: number,
-  targetListId: string,
-  targetIndex: number,
-}
-
-type KanbanAction = KanbanAddItemAction | KanbanMoveItemAction | KanbanUpdateItemAction;
 
 // https://github.com/DefinitelyTyped/DefinitelyTyped/pull/24509
 export const KanbanActionDispatchContext = React.createContext<Dispatch<KanbanAction>>(undefined as never);
 
-function stateReducer(state: KanbanState, action: KanbanAction): KanbanState {
-  // the following deep copy is slow, but I would use 'Structured Cloning' in an up-to-date NodeJS:
-  // https://stackoverflow.com/questions/122102/what-is-the-most-efficient-way-to-deep-clone-an-object-in-javascript
-  // Naturally, merging the state with the changes to preserve as much state as possible
-  // would be more efficient but more work.
-  const newState: KanbanState = JSON.parse(JSON.stringify(state));
-  switch (action.type) {
-    case "addItem": {
-      const relevantList = newState.lists.find(value => value.id === action.targetListId);
-      assert(relevantList !== undefined);
-      relevantList.items.push(action.item);
-      return newState;
-    }
-    case "moveItem": {
-      const sourceList = newState.lists.find(value => value.id === action.sourceListId);
-      assert(sourceList !== undefined);
-      const targetList = newState.lists.find(value => value.id === action.targetListId);
-      assert(targetList !== undefined);
-
-      const item = sourceList.items[action.sourceIndex];
-      assert(item.id === action.itemId);
-
-      // remove item first
-      sourceList.items.splice(action.sourceIndex, 1);
-
-      // then add it to the target list
-      targetList.items.splice(action.targetIndex, 0, item);
-
-      return newState;
-    }
-    case "updateItem": {
-      // note: super inefficient, but whatever for this example.
-      const relevantList = newState.lists.find(value => value.items.find(item => item.id === action.itemId) !== undefined);
-      assert(relevantList !== undefined);
-
-      const item = relevantList.items.find(item => item.id === action.itemId);
-      assert(item !== undefined);
-      assert(item.id === action.itemId);
-
-      item.checked = action.checked;
-
-      return newState;
-    }
-    default:
-      throw new Error();
-  }
+// TODO how to nicely tell typescript that an emtpy object will be passed as arg?
+// eslint-disable-next-line
+export function Kanban({}: any) {
+  const [state, dispatch] = useReducer(stateReducer, null, () => createInitialState());
+  return (
+    <Paper sx={{paddingBottom: 4}}>
+      <KanbanActionDispatchContext.Provider value={dispatch}>
+        <KanbanDragAndDropContext>
+          <KanbanListsComponent lists={state.lists}/>
+        </KanbanDragAndDropContext>
+      </KanbanActionDispatchContext.Provider>
+    </Paper>
+  );
 }
 
 export function KanbanDragAndDropContext({children}: { children: ReactChildOrChildren }) {
@@ -143,21 +88,6 @@ export function KanbanDragAndDropContext({children}: { children: ReactChildOrChi
   );
 }
 
-// TODO how to nicely tell typescript that an emtpy object will be passed as arg?
-// eslint-disable-next-line
-export function Kanban({}: any) {
-  const [state, dispatch] = useReducer(stateReducer, null, () => createInitialState());
-  return (
-    <Paper sx={{paddingBottom: 4}}>
-      <KanbanActionDispatchContext.Provider value={dispatch}>
-        <KanbanDragAndDropContext>
-          <KanbanListsComponent lists={state.lists}/>
-        </KanbanDragAndDropContext>
-      </KanbanActionDispatchContext.Provider>
-    </Paper>
-  );
-}
-
 function KanbanListsComponent({lists}: { lists: KanbanList[] }) {
   return (
     <Stack spacing={2} margin={5} direction="row">
@@ -165,58 +95,6 @@ function KanbanListsComponent({lists}: { lists: KanbanList[] }) {
         return <KanbanListComponent key={list.id} list={list}/>
       })}
     </Stack>
-  );
-}
-
-function getKanbanItemStyle(style: DraggingStyle | NotDraggingStyle | undefined,
-                            snapshot: DraggableStateSnapshot): CSSProperties {
-  if (snapshot.isDropAnimating) {
-    // https://github.com/atlassian/react-beautiful-dnd/blob/master/docs/guides/drop-animation.md
-    const notDraggingStyle = style as NotDraggingStyle;
-    return {
-      ...notDraggingStyle,
-      // Apparently it is important, that this is the last property so that it overrides the 'transition' aspect.
-      transitionDuration: `0.15s`, // default is 0.34s
-    };
-  } else if (snapshot.isDragging) {
-    const draggingStyle = style as DraggingStyle;
-    draggingStyle.opacity = 0.8;
-    return draggingStyle as CSSProperties;
-  } else {
-    return style as CSSProperties;
-  }
-}
-
-function KanbanItemComponent({item, index}: { item: KanbanItem, index: number }) {
-  const dispatch: Dispatch<KanbanAction> = useContext(KanbanActionDispatchContext);
-
-  function onCheckedChanged(event: React.ChangeEvent<HTMLInputElement>) {
-    dispatch({
-      type: 'updateItem',
-      itemId: item.id,
-      checked: event.target.checked,
-    });
-  }
-
-  return (
-    <Draggable draggableId={item.id} index={index}>
-      {(provided, snapshot) => (
-        <div ref={provided.innerRef}
-             {...provided.draggableProps}
-             {...provided.dragHandleProps}
-             style={getKanbanItemStyle(provided.draggableProps.style, snapshot)}>
-          <Card>
-            <CardContent>
-              <Stack spacing={2} direction="row" alignItems="center">
-                <Checkbox checked={item.checked} onChange={onCheckedChanged}/>
-                <Typography variant="h6">{item.content}</Typography>
-              </Stack>
-            </CardContent>
-          </Card>
-        </div>
-      )
-      }
-    </Draggable>
   );
 }
 
@@ -264,6 +142,58 @@ export function KanbanListComponent({list}: { list: KanbanList }) {
       }}
     </Droppable>
   );
+}
+
+function KanbanItemComponent({item, index}: { item: KanbanItem, index: number }) {
+  const dispatch: Dispatch<KanbanAction> = useContext(KanbanActionDispatchContext);
+
+  function onCheckedChanged(event: React.ChangeEvent<HTMLInputElement>) {
+    dispatch({
+      type: 'updateItem',
+      itemId: item.id,
+      checked: event.target.checked,
+    });
+  }
+
+  return (
+    <Draggable draggableId={item.id} index={index}>
+      {(provided, snapshot) => (
+        <div ref={provided.innerRef}
+             {...provided.draggableProps}
+             {...provided.dragHandleProps}
+             style={getKanbanItemStyle(provided.draggableProps.style, snapshot)}>
+          <Card>
+            <CardContent>
+              <Stack spacing={2} direction="row" alignItems="center">
+                <Checkbox checked={item.checked} onChange={onCheckedChanged}/>
+                <Typography variant="h6">{item.content}</Typography>
+              </Stack>
+            </CardContent>
+          </Card>
+        </div>
+      )
+      }
+    </Draggable>
+  );
+}
+
+function getKanbanItemStyle(style: DraggingStyle | NotDraggingStyle | undefined,
+                            snapshot: DraggableStateSnapshot): CSSProperties {
+  if (snapshot.isDropAnimating) {
+    // https://github.com/atlassian/react-beautiful-dnd/blob/master/docs/guides/drop-animation.md
+    const notDraggingStyle = style as NotDraggingStyle;
+    return {
+      ...notDraggingStyle,
+      // Apparently it is important, that this is the last property so that it overrides the 'transition' aspect.
+      transitionDuration: `0.15s`, // default is 0.34s
+    };
+  } else if (snapshot.isDragging) {
+    const draggingStyle = style as DraggingStyle;
+    draggingStyle.opacity = 0.8;
+    return draggingStyle as CSSProperties;
+  } else {
+    return style as CSSProperties;
+  }
 }
 
 export default Kanban;
